src/requirements.txt
django==1.9.3
psycopg2==2.6.1
gunicorn==19.4.5
python-dateutil==2.5.0
inflect==0.2.5
